+++ examples/build-spf.rs
@@ -4,12 +4,15 @@ use decon_spf::Spf;
 fn main() {
     let mut spf1 = Spf::new();
     spf1.set_v1();
-    let ip_m_1 = ParsedMechanism::new("+ip4:203.32.160.0/24").unwrap();
-    let ip_m_2 = ParsedMechanism::new("+ip4:203.32.166.0/24").unwrap();
+    let ip_m_1 = ParsedMechanism::new("ip4:203.32.160.0/24");
+    let ip_m_2 = ParsedMechanism::new("+ip4:203.32.166.0/24");
     let mx = ParsedMechanism::new("mx").unwrap();
-
-    spf1.append_ip_mechanism(ip_m_1.network());
-    spf1.append_ip_mechanism(ip_m_2.network());
+    if let Ok(ip1) = ip_m_1 {
+        spf1.append_ip_mechanism(ip1.network());
+    }
+    if let Ok(ip2) = ip_m_2 {
+        spf1.append_ip_mechanism(ip2.network());
+    }
     spf1.append_mechanism(mx.txt());
 
     spf1.append_mechanism("a:test.com".parse().unwrap());